refactor(amdsmi-tests): split CLI tests into per-command modules
Replace the four device-grouped cli/test_cli_{common,gpu,cpu,topology}.py
files with one module per amd-smi command (test_<command>.py, class
Test<Command>, primary method test_command). Move the shared --json
baseline (metric/static/list/partition data, the derived gpus list and
sub_args dict) into a cached TestCliBase.setUpClass so it runs once for
the whole CLI suite instead of once per group, and resolve the cached
attributes via normal class inheritance rather than dynamic setattr.

Update test-design.md and the test-id example comment to match.
