+++ projects/amdsmi/tests/python/cli/test_event.py
@@ -1,0 +1,45 @@
+#!/usr/bin/env python3
+#
+# Copyright (C) Advanced Micro Devices. All rights reserved.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy of
+# this software and associated documentation files (the "Software"), to deal in
+# the Software without restriction, including without limitation the rights to
+# use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+# the Software, and to permit persons to whom the Software is furnished to do so,
+# subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+# FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+# COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+# IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+# CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+"""CLI leaf test: event command."""
+
+from cli.base import TestCliBase
+
+
+class TestEvent(TestCliBase):
+    def test_command(self):
+        self.common.print_func_name("")
+        msg = f"{self.tab}### amd-smi event"
+        self.common.print(msg)
+
+        # TODO allow event commands to be executed
+        if not self.PrintCmdsOnly:
+            if self.common.TODO_SKIP_FAIL:
+                msg = f"{self.tab}Needs input"
+                self.common.print(msg)
+                self.skipTest(msg)
+
+        # Start process with "amd-smi event"
+        # In another process create an event with like "amd-smi reset --gpureset"
+        cmds = self.CreateCmds(
+            "event", "Event Arguments:", "Device Arguments:", "Command Modifiers:", ""
+        )
+        self.RunCmds(cmds)
+        return
